Move private members into BiCouplingScheme
* Move _m2n, _sendData, _receiveData into BiCouplingScheme, since these private members are not needed/used by MultiCouplingScheme
* Move functions that depend on these members into BiCouplingScheme, since these functions are not needed/used by MultiCouplingScheme, as well
* Use BiCouplingScheme in configuration, if appropriate
